Fix details of balance
