+++ src/tenant/balance/infrastructure/data-provider/indexedDB/adapter/balance.dexie.adapter.indexedDB.data-provider.ts
@@ -7,7 +7,7 @@ import {IBalance} from "@tenant/balance/domain";
 @Injectable()
 export class BalanceDexieAdapterIndexedDBDataProvider extends DexieAdapterIndexedDBDataProvider<IBalance.EntityRaw> {
 
-	protected readonly columns = '_id,state,createdAt,updatedAt';
+	protected readonly columns = '_id,state,createdAt,updatedAt,action.amount,action.type,action.systemComment,action.currency,amountAfterAction,amountBeforeAction';
 	protected readonly moduleName = 'balance';
 	protected readonly version = 1;
