+++ src/tenant/balance/presentation/ui/component/details/balance-details-container.component.ts
@@ -1,13 +1,9 @@
-import {Component, inject, input, ViewEncapsulation} from '@angular/core';
+import {Component, input, ViewEncapsulation} from '@angular/core';
 import {TranslateModule} from "@ngx-translate/core";
-import {ActiveStyleDirective} from "@shared/presentation/directives/active-style/active-style.directive";
-import {PrimaryLinkStyleDirective} from "@shared/presentation/directives/link/primary.link.style.directive";
-import {
-	RowActionButtonComponent
-} from "@tenant/customer/presentation/ui/component/row-action-button/row-action-button.component";
-import ECustomer from "@tenant/customer/domain/entity/e.customer";
-import {Router} from "@angular/router";
 import {StandardDetailsEntityComponent} from "@shared/presentation/component/entity/standard-details.entity.component";
+import EBalance from "@tenant/balance/domain/entity/e.balance";
+import {BalanceActionTypeEnum} from "@tenant/balance/domain";
+import {CurrencyPipe} from "@angular/common";
 
 @Component({
 	selector: 'balance-detail-page',
@@ -15,10 +11,8 @@ import {StandardDetailsEntityComponent} from "@shared/presentation/component/ent
 	encapsulation: ViewEncapsulation.None,
 	imports: [
 		TranslateModule,
-		ActiveStyleDirective,
-		PrimaryLinkStyleDirective,
-		RowActionButtonComponent,
-		StandardDetailsEntityComponent
+		StandardDetailsEntityComponent,
+		CurrencyPipe,
 	],
 	host: {
 		class: 'bg-neutral-100'
@@ -29,19 +23,9 @@ export class BalanceDetailsContainerComponent {
 
 	// TODO add base index of details with store and delete method
 
-	public readonly item = input.required<ECustomer>();
+	public readonly item = input.required<EBalance>();
 
-	private readonly router = inject(Router);
-
-
-	public async openCustomersOrders() {
-		const customer = this.item();
-		if (customer) {
-			await this.router.navigate([{outlets: {second: ['balance', customer._id, 'order']}}], {
-				onSameUrlNavigation: 'reload',
-			});
-		}
-	}
+	protected readonly balanceActionTypeEnum = BalanceActionTypeEnum;
 }
 
 export default BalanceDetailsContainerComponent;
